assign dummy id to response

+++ models/chat.go
@@ -1,8 +1,8 @@
 package models
 
 type Chat struct {
-	Id        int   `json:"id"`
-	UserId    int   `json:"user_id"`
+	Id        int64 `json:"id"`
+	UserId    int64 `json:"user_id"`
 	CreatedAt int64 `json:"created_at"`
 	DeletedAt int64 `json:"deleted_at"`
 }
