+++ models/request.go
@@ -1,14 +1,14 @@
 package models
 
 type Request struct {
-	Id                int        `json:"id"`
-	UserId            int        `json:"user_id"`
-	ChatId            int        `json:"chat_id"`
+	Id                int64      `json:"id"`
+	UserId            int64      `json:"user_id"`
+	ChatId            int64      `json:"chat_id"`
 	StartedAt         int64      `json:"started_at"`
 	FinishedAt        int64      `json:"finished_at"`
 	Latency           int64      `json:"latency"`
-	Chunks            int        `json:"chunks"`
-	Attempts          int        `json:"attempts"`
+	Chunks            int64      `json:"chunks"`
+	Attempts          int64      `json:"attempts"`
 	Language          string     `json:"language"`
 	SystemInstruction string     `json:"system_instruction"`
 	Contents          []*Message `json:"contents"`
@@ -17,7 +17,7 @@ type Request struct {
 	ToolCalls         []string   `json:"tool_calls"`
 	FinishReason      string     `json:"finish_reason"`
 	Model             string     `json:"model"`
-	PromptTokens      int        `json:"prompt_tokens"`
-	ResponseTokens    int        `json:"response_tokens"`
+	PromptTokens      int64      `json:"prompt_tokens"`
+	ResponseTokens    int64      `json:"response_tokens"`
 	Price             float64    `json:"price"`
 }
